Make the notebook execution buffer move repeatedly executed cells to the end of the queue.
This is consistent with how execution works, since the cell is only finally executed with the final text of the cell.

This may still give surprising results, for example, if cell 1 defines a variable and cell 2 increments and prints the variable.
If the execution order is cell 1, cell 2, cell 1, then before this commit cell 2 would print fine, whereas now cell 2 will throw an error.